requirements.txt
Flask>=2.3
gunicorn>=21.2
psycopg2-binary>=2.9
flask-cors>=4.0
